Rename facebook file
